Added support for python 3.5
It replaces the use of attrfind in favor of attrfind_tolerant, as
the former was removed in python 3.5.

For python 2 and pypy, we fallback to attrfind as attrfind_tolerant is not
available.

For python 3.4 and 3.5 we set convert_charrefs to False. Python 3.5
sets its by default to True, which does not allow to introspect their value.

Fixes deezer/html-linter#8.